requirements.txt
python-telegram-bot[speedups]==20.5
flask==3.0.2
requests==2.31.0
python-dotenv==1.0.0
sqlite3==2.6.0